Refactor table.rs into query/mod.rs module
