+++ server/src/query/mod.rs
@@ -13,52 +13,47 @@
  * limitations under the License.
  */
 
-//! Implementation of the types required to query model tables through Apache
-//! Arrow DataFusion. The types are [`ModelTable`] which implements
-//! [`TableProvider`], [`GridExec`] which implements [`ExecutionPlan`], and
-//! [`GridStream`] which implements [`Stream`] and [`RecordBatchStream`].
+//! Implementation of the types required to query model tables through Apache Arrow DataFusion. The
+//! types are:
+//! * [`ModelTable`] which takes the projection, filters as [`Exprs`](Expr), and limit as input and
+//! produces the part of a physical query plan that produces the data points required for the query.
+//! * [`GridExec`] which takes the sorted compressed segments from the Apache Parquet files for a
+//! single column and reconstructs the data points they represent as three sorted arrays containing
+//! the data points' univariate ids, timestamps, and values.
+//! * `SortedJoinExec` which take the sorted arrays produced by each [`GridExec`] and combines them
+//! with the time series tags retrieved from the
+//! [`MetadataManager`](crate::metadata::MetadataManager) to create the complete results containing
+//! a timestamp column, one or more field columns, and zero or more tag columns.
+// TODO: implemented SortedJoinExec and make its `` a link.
+// TODO: determine if ParquetExec filtering is enough and the extra FilterExec can be removed.
+
+// Public so the rules added to Apache Arrow DataFusion's physical optimizer can access GridExec.
+pub mod grid_exec;
 
 use std::any::Any;
-use std::collections::HashMap;
-use std::fmt;
-use std::fmt::Formatter;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::{Context as StdTaskContext, Poll};
 
 use async_trait::async_trait;
-use datafusion::arrow::array::{
-    ArrayRef, BinaryArray, Float32Array, StringBuilder, UInt64Array, UInt8Array,
-};
 use datafusion::arrow::datatypes::{ArrowPrimitiveType, Field, Schema, SchemaRef};
-use datafusion::arrow::error::Result as ArrowResult;
-use datafusion::arrow::record_batch::RecordBatch;
 use datafusion::common::ToDFSchema;
 use datafusion::datasource::object_store::ObjectStoreUrl;
 use datafusion::datasource::{
     datasource::TableProviderFilterPushDown, listing::PartitionedFile, TableProvider, TableType,
 };
 use datafusion::error::{DataFusionError, Result};
-use datafusion::execution::context::{ExecutionProps, SessionState, TaskContext};
+use datafusion::execution::context::{ExecutionProps, SessionState};
 use datafusion::logical_expr::{self, BinaryExpr, Expr, Operator};
 use datafusion::optimizer::utils;
 use datafusion::physical_expr::planner;
 use datafusion::physical_plan::{
-    expressions::PhysicalSortExpr, file_format::FileScanConfig, file_format::ParquetExec,
-    filter::FilterExec, metrics::BaselineMetrics, metrics::ExecutionPlanMetricsSet,
-    metrics::MetricsSet, DisplayFormatType, ExecutionPlan, Partitioning, RecordBatchStream,
-    SendableRecordBatchStream, Statistics,
+    file_format::FileScanConfig, file_format::ParquetExec, filter::FilterExec, ExecutionPlan,
+    Statistics,
 };
-use futures::stream::{Stream, StreamExt};
 
 use crate::metadata::model_table_metadata::ModelTableMetadata;
-use crate::metadata::MetadataManager;
-use crate::models;
+use crate::query::grid_exec::GridExec;
 use crate::storage;
-use crate::types::{
-    ArrowTimestamp, ArrowUnivariateId, ArrowValue, CompressedSchema, TimestampArray,
-    TimestampBuilder, ValueArray, ValueBuilder,
-};
+use crate::types::{ArrowTimestamp, ArrowUnivariateId, ArrowValue, CompressedSchema};
 use crate::Context;
 
 /// A queryable representation of a model table which stores multivariate time
@@ -394,318 +389,6 @@ impl TableProvider for ModelTable {
         );
 
         Ok(grid_exec)
-    }
-}
-
-/// An operator that reconstructs the data points stored as segments containing
-/// metadata and models. It is public so the additional rules added to Apache
-/// Arrow DataFusion's physical optimizer can pattern match on it.
-#[derive(Debug, Clone)]
-pub struct GridExec {
-    /// Metadata required to query the model table.
-    model_table_metadata: Arc<ModelTableMetadata>,
-    /// Columns requested by the query.
-    projection: Vec<usize>,
-    /// Number of rows requested by the query.
-    limit: Option<usize>,
-    /// Schema of the model table after projection.
-    schema_after_projection: SchemaRef,
-    /// Mapping from tag hash to tags.
-    hash_to_tags: HashMap<u64, Vec<String>>,
-    /// Operator to read batches of rows from.
-    input: Arc<dyn ExecutionPlan>,
-    /// Metrics collected during execution for use by EXPLAIN ANALYZE.
-    metrics: ExecutionPlanMetricsSet,
-}
-
-impl GridExec {
-    pub fn new(
-        model_table_metadata: Arc<ModelTableMetadata>,
-        projection: Vec<usize>,
-        limit: Option<usize>,
-        schema: SchemaRef,
-        hash_to_tags: HashMap<u64, Vec<String>>,
-        input: Arc<dyn ExecutionPlan>,
-    ) -> Arc<Self> {
-        // Modifies the schema so it matches the passed projection. unwrap() is safe as the
-        // projection is for the schema so errors due to out of bounds indexing cannot happen.
-        let schema_after_projection = Arc::new(schema.project(&projection).unwrap());
-
-        Arc::new(GridExec {
-            model_table_metadata,
-            projection,
-            limit,
-            schema_after_projection,
-            hash_to_tags,
-            input,
-            metrics: ExecutionPlanMetricsSet::new(),
-        })
-    }
-}
-
-#[async_trait]
-impl ExecutionPlan for GridExec {
-    /// Return `self` as [`Any`] so it can be downcast.
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    /// Return the schema of the model table after projection.
-    fn schema(&self) -> SchemaRef {
-        self.schema_after_projection.clone()
-    }
-
-    /// Return the single operator batches of rows are read from.
-    fn children(&self) -> Vec<Arc<dyn ExecutionPlan>> {
-        vec![self.input.clone()]
-    }
-
-    /// Return the partitioning of the single operator batches of rows are read
-    /// from as [`GridExec`] does not repartition the batches of rows.
-    fn output_partitioning(&self) -> Partitioning {
-        self.input.output_partitioning()
-    }
-
-    /// Return `None` to indicate that `GridExec` does not guarantee a specific
-    /// ordering of the rows it produces.
-    fn output_ordering(&self) -> Option<&[PhysicalSortExpr]> {
-        //TODO: can it be guaranteed that GridExec outputs ordered data?
-        None
-    }
-
-    /// Return a new instance of [`GridExec`] with the operator to read batches
-    /// of rows from replaced. [`DataFusionError::Plan`] is returned if
-    /// `children` does not contain a single element.
-    fn with_new_children(
-        self: Arc<Self>,
-        children: Vec<Arc<dyn ExecutionPlan>>,
-    ) -> Result<Arc<dyn ExecutionPlan>> {
-        if children.len() == 1 {
-            Ok(Arc::new(GridExec {
-                model_table_metadata: self.model_table_metadata.clone(),
-                projection: self.projection.clone(),
-                limit: self.limit,
-                schema_after_projection: self.schema_after_projection.clone(),
-                hash_to_tags: self.hash_to_tags.clone(),
-                input: children[0].clone(),
-                metrics: self.metrics.clone(),
-            }))
-        } else {
-            Err(DataFusionError::Plan(format!(
-                "A single child must be provided {self:?}"
-            )))
-        }
-    }
-
-    /// Create a stream that read batches of rows with segments from the data
-    /// source operator, reconstructs the data points from the metadata and
-    /// models in the segments, and returns batches of rows with data points.
-    fn execute(
-        &self,
-        partition: usize,
-        task_context: Arc<TaskContext>,
-    ) -> Result<SendableRecordBatchStream> {
-        Ok(Box::pin(GridStream::new(
-            self.projection.clone(),
-            self.limit,
-            self.schema_after_projection.clone(),
-            self.hash_to_tags.clone(), // TODO: share single instance though RC or ARC.
-            self.input.execute(partition, task_context)?,
-            BaselineMetrics::new(&self.metrics, partition),
-        )))
-    }
-
-    /// Specify that [`GridExec`] knows nothing about the data it will output.
-    fn statistics(&self) -> Statistics {
-        Statistics {
-            num_rows: None,
-            total_byte_size: None,
-            column_statistics: None,
-            is_exact: false,
-        }
-    }
-
-    /// Return a snapshot of the set of metrics being collected by the operator.
-    fn metrics(&self) -> Option<MetricsSet> {
-        Some(self.metrics.clone_inner())
-    }
-
-    /// Write a string-based representation of the operator to `f`. Returns
-    /// `Err` if `std::write` cannot format the string and write it to `f`.
-    fn fmt_as(&self, _t: DisplayFormatType, f: &mut Formatter<'_>) -> fmt::Result {
-        let columns: Vec<&String> = self
-            .schema_after_projection
-            .fields()
-            .iter()
-            .map(|f| f.name())
-            .collect();
-
-        write!(
-            f,
-            "GridExec: projection={:?}, limit={:?}, columns={:?}",
-            self.projection, self.limit, columns
-        )
-    }
-}
-
-/// A stream that read batches of rows with segments from the data source
-/// operator, reconstructs the data points from the metadata and models in the
-/// segments, and returns batches of rows with data points.
-struct GridStream {
-    /// Columns requested by the query.
-    projection: Vec<usize>,
-    /// Number of rows requested by the query.
-    _limit: Option<usize>,
-    /// Schema of the model table after projection.
-    schema_after_projection: SchemaRef,
-    /// Mapping from tag hash to tags.
-    hash_to_tags: HashMap<u64, Vec<String>>,
-    /// Stream to read batches of rows from.
-    input: SendableRecordBatchStream,
-    /// Metrics collected during execution for use by EXPLAIN ANALYZE.
-    baseline_metrics: BaselineMetrics,
-}
-
-impl GridStream {
-    fn new(
-        projection: Vec<usize>,
-        limit: Option<usize>,
-        schema_after_projection: SchemaRef,
-        hash_to_tags: HashMap<u64, Vec<String>>,
-        input: SendableRecordBatchStream,
-        baseline_metrics: BaselineMetrics,
-    ) -> Self {
-        GridStream {
-            projection,
-            _limit: limit,
-            schema_after_projection,
-            hash_to_tags,
-            input,
-            baseline_metrics,
-        }
-    }
-
-    // TODO: it is necessary to return batch_size data points to prevent skew?
-    // TODO: limit the batches of data points to only contain what is needed.
-    /// Reconstruct the data points from the metadata and models in the segments
-    /// in `batch`, and return batches of rows with data points.
-    fn grid(&self, batch: &RecordBatch) -> RecordBatch {
-        // Record the time elapsed from the timer is created to it is dropped.
-        let _timer = self.baseline_metrics.elapsed_compute().timer();
-
-        // Retrieve the arrays from batch and cast them to their concrete type.
-        crate::arrays!(
-            batch,
-            univariate_ids,
-            model_type_ids,
-            start_times,
-            end_times,
-            timestamps,
-            min_values,
-            max_values,
-            values,
-            _error_array
-        );
-
-        // Each segment is guaranteed to contain at least one data point.
-        let num_rows = batch.num_rows();
-        let mut univariate_ids_builder = UInt64Array::builder(num_rows);
-        let mut timestamp_builder = TimestampBuilder::with_capacity(num_rows);
-        let mut value_builder = ValueBuilder::with_capacity(num_rows);
-
-        // Reconstructs the data points from the segments.
-        for row_index in 0..num_rows {
-            // unwrap() is safe as the storage engine created the strings.
-            let univariate_id = univariate_ids.value(row_index);
-            let model_type_id = model_type_ids.value(row_index);
-            let start_time = start_times.value(row_index);
-            let end_time = end_times.value(row_index);
-            let timestamps = timestamps.value(row_index);
-            let min_value = min_values.value(row_index);
-            let max_value = max_values.value(row_index);
-            let values = values.value(row_index);
-
-            models::grid(
-                univariate_id,
-                model_type_id,
-                timestamps,
-                start_time,
-                end_time,
-                values,
-                min_value,
-                max_value,
-                &mut univariate_ids_builder,
-                &mut timestamp_builder,
-                &mut value_builder,
-            );
-        }
-
-        // Append the tags to each of the data points if any were requested.
-        let mut tag_columns: Vec<StringBuilder> = Vec::with_capacity(self.projection.len());
-
-        // Skip appending tags if none were requested.
-        if !self.hash_to_tags.is_empty() {
-            // unwrap() is safe as the HashMap is guaranteed to contain at least one value.
-            let tag_columns_len = self.hash_to_tags.values().next().unwrap().len();
-            for _ in 0..tag_columns_len {
-                tag_columns.push(StringBuilder::new());
-            }
-
-            for univariate_id in univariate_ids_builder.values_slice() {
-                let tag_hash = MetadataManager::univariate_id_to_tag_hash(*univariate_id);
-                let tags = &self.hash_to_tags[&tag_hash];
-                for index in 0..tag_columns_len {
-                    tag_columns[index].append_value(tags[index].clone());
-                }
-            }
-        }
-
-        // Add the none tag columns to the batch.
-        let mut columns: Vec<ArrayRef> = Vec::with_capacity(self.projection.len());
-        for column in &self.projection {
-            match column {
-                0 => columns.push(Arc::new(univariate_ids_builder.finish())),
-                1 => columns.push(Arc::new(timestamp_builder.finish())),
-                2 => columns.push(Arc::new(value_builder.finish())),
-                _ => (), // The remaining columns are for tags.
-            }
-        }
-
-        // Add the tag columns to the batch.
-        tag_columns
-            .drain(0..)
-            .for_each(|mut tag_column| columns.push(Arc::new(tag_column.finish())));
-
-        // Return the batch, unwrap() is safe as columns are constructed from self.projection.
-        RecordBatch::try_new(self.schema_after_projection.clone(), columns).unwrap()
-    }
-}
-
-impl Stream for GridStream {
-    /// Specify that [`GridStream`] returns [`ArrowResult<RecordBatch>`] when
-    /// polled.
-    type Item = ArrowResult<RecordBatch>;
-
-    /// Try to poll the next element from the [`GridStream`] and returns:
-    /// * `Poll::Pending` if the next element is not yet ready.
-    /// * `Poll::Ready(Some(Ok(batch)))` if an element is ready.
-    /// * `Poll::Ready(None)` if the stream is empty.
-    fn poll_next(
-        mut self: Pin<&mut Self>,
-        cx: &mut StdTaskContext<'_>,
-    ) -> Poll<Option<Self::Item>> {
-        let poll = self.input.poll_next_unpin(cx).map(|x| match x {
-            Some(Ok(batch)) => Some(Ok(self.grid(&batch))),
-            other => other,
-        });
-        self.baseline_metrics.record_poll(poll)
-    }
-}
-
-impl RecordBatchStream for GridStream {
-    /// Return the schema of the model table after projection.
-    fn schema(&self) -> SchemaRef {
-        self.schema_after_projection.clone()
     }
 }
